feat: refactor keypoint handling and add utility functions for single view annotation. To ensure beartype runtime type checking works
