+++ annotation_example/__init__.py
@@ -4,6 +4,9 @@ from beartype.claw import beartype_this_package
 
 beartype_this_package(
     conf=BeartypeConf(
-        claw_skip_package_names=("annotation_example.gradio_ui.mv_sam_callbacks", "annotation_example.gradio_ui.sv_sam")
+        claw_skip_package_names=(
+            "annotation_example.gradio_ui.mv_sam_callbacks",
+            "annotation_example.gradio_ui.sv_sam_callbacks",
+        )
     )
 )
