+++ annotation_example/gradio_ui/sv_sam_callbacks.py
@@ -1,11 +1,12 @@
 import uuid
 from pathlib import Path
-from typing import Literal
+from typing import Literal, TypedDict
 
 import gradio as gr
 import rerun as rr
-from gradio_rerun.events import SelectionChange
-from typing_extensions import TypedDict
+from gradio_rerun.events import (
+    SelectionChange,
+)
 
 from annotation_example.gradio_ui.utils import KeypointsContainer, get_recording
 
@@ -13,24 +14,22 @@ from annotation_example.gradio_ui.utils import KeypointsContainer, get_recording
 class RerunLogPaths(TypedDict):
     timeline_name: str
     parent_log_path: Path
-    cam_log_path_list: list[Path]
+    camera_log_path: Path
+    pinhole_path: Path
 
 
 # In this function, the `request` and `evt` parameters will be automatically injected by Gradio when this event listener is fired.
 #
 # `SelectionChange` is a subclass of `EventData`: https://www.gradio.app/docs/gradio/eventdata
 # `gr.Request`: https://www.gradio.app/main/docs/gradio/request
-def update_keypoints(
+def single_view_update_keypoints(
     active_recording_id: uuid.UUID,
     point_type: Literal["include", "exclude"],
-    mv_keypoint_dict: dict[str, KeypointsContainer],
+    keypoints_container: KeypointsContainer,
     log_paths: RerunLogPaths,
     request: gr.Request,
     change: SelectionChange,
 ):
-    if active_recording_id == "":
-        return
-
     evt = change.payload
 
     # We can only log a keypoint if the user selected only a single item.
@@ -46,28 +45,22 @@ def update_keypoints(
     rec: rr.RecordingStream = get_recording(active_recording_id)
     stream: rr.BinaryStream = rec.binary_stream()
     current_keypoint: tuple[int, int] = tuple(item.position[0:2])
+    keypoints_container.add_point(current_keypoint, point_type)
 
-    for cam_name in mv_keypoint_dict:
-        if cam_name in item.entity_path:
-            # Update the keypoints for the specific camera
-            mv_keypoint_dict[cam_name].add_point(current_keypoint, point_type)
-            current_keypoint_container: KeypointsContainer = mv_keypoint_dict[cam_name]
-
-    rec.set_time_nanos(log_paths["timeline_name"], nanos=0)
+    rec.set_time_sequence(log_paths["timeline_name"], sequence=0)
     # Log include points if any exist
-    if current_keypoint_container.include_points.shape[0] > 0:
+    if keypoints_container.include_points.shape[0] > 0:
         rec.log(
-            f"{item.entity_path}/include",
-            rr.Points2D(current_keypoint_container.include_points, colors=(0, 255, 0), radii=5),
+            f"{item.entity_path}/include", rr.Points2D(keypoints_container.include_points, colors=(0, 255, 0), radii=5)
         )
 
     # Log exclude points if any exist
-    if current_keypoint_container.exclude_points.shape[0] > 0:
+    if keypoints_container.exclude_points.shape[0] > 0:
         rec.log(
             f"{item.entity_path}/exclude",
-            rr.Points2D(current_keypoint_container.exclude_points, colors=(255, 0, 0), radii=5),
+            rr.Points2D(keypoints_container.exclude_points, colors=(255, 0, 0), radii=5),
         )
 
-    # # Ensure we consume everything from the recording.
+    # Ensure we consume everything from the recording.
     stream.flush()
-    yield stream.read(), mv_keypoint_dict
+    yield stream.read(), keypoints_container
